HW Работа с файлами

+++ src/main/java/ru/hogwarts/school/service/FacultyService.java
@@ -6,6 +6,7 @@ import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.repositories.FacultyRepository;
 
 import java.util.Collection;
+import java.util.Optional;
 
 @Service
 public class FacultyService {
@@ -14,6 +15,7 @@ public class FacultyService {
     public FacultyService(FacultyRepository facultyRepository) {
         this.facultyRepository = facultyRepository;
     }
+
     public Collection<Student> getFacultyStudents(long facultyId) {
         Faculty faculty = facultyRepository.findById(facultyId).orElse(null);
         if (faculty != null) {
@@ -21,14 +23,15 @@ public class FacultyService {
         }
         return null;
     }
+
     public Faculty createFaculty(Faculty faculty) {
 
         return facultyRepository.save(faculty);
     }
 
     public Faculty findFaculty(long id) {
-
-        return facultyRepository.findById(id).get();
+        Optional<Faculty> facultyOptional = facultyRepository.findById(id);
+        return facultyOptional.orElse(null);
     }
 
     public Faculty editFaculty(Faculty updatedFaculty) {
@@ -47,15 +50,21 @@ public class FacultyService {
         }
     }
 
-    public void deleteFaculty(long id) {
-
-        facultyRepository.deleteById(id);
+    public Faculty deleteFaculty(long id) {
+        Optional<Faculty> facultyOptional = facultyRepository.findById(id);
+        if (facultyOptional.isPresent()) {
+            facultyRepository.deleteById(id);
+            return facultyOptional.get();
+        } else {
+            return null;
+        }
     }
 
     public Collection<Faculty> getAllFaculties() {
 
         return facultyRepository.findAll();
     }
+
     public Collection<Faculty> findFacultiesByNameOrColorIgnoreCase(String query) {
         return facultyRepository.findFacultiesByNameIgnoreCaseOrColorIgnoreCase(query, query);
     }
